% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grafico.R
\name{grafico_pred_obs}
\alias{grafico_pred_obs}
\title{Regressão Linear Múltipla}
\usage{
grafico_pred_obs(regressao)
}
\arguments{
\item{regressao}{Um modelo de regressão ajustado com a função \code{regressao()}.}
}
\value{
Um gráfico de com valores preditos no eixo x e valores observados no eixo y.
}
\description{
Esta função retorna um gráfico de valores preditos \eqn{\times} valores observados como base em modelo de regressão linear ajustado.
}
\examples{
ajuste1 <- regressao(X = c('B'), Y = 'A', dados = sodad)
grafico_pred_obs(ajuste1)
ajuste2 <- regressao(X = c('A', 'B'), Y = 'C', dados = sodad)
grafico_pred_obs(ajuste2)
ajuste3 <- regressao(X = c('A', 'B', 'D'), Y = 'E', dados = sodad)
grafico_pred_obs(ajuste3)
ajuste4 <- regressao(X = c('A', 'C', 'D', 'E'), Y = 'B', dados = sodad)
grafico_pred_obs(ajuste4)

}
